Adds shell script to run flake8.
This flake8 command is actually defined in setup.cfg, but that's not
obvious. It being defined there is how E266 is ignored. It could
also be used to have rules ignored for specific files as well.
Having a shell script that explains this is the best way to make this
information available to others.